laba5/requirements.txt
requests
numpy
pandas
matplotlib
Pillow
faker
beautifulsoup4
lxml
sympy
seaborn
